Refactor and Modularize data pipeline script
