--- example1/my_second_dbt_model.sql
@@ -1,6 +1,0 @@
-
--- Use the `ref` function to select from other models
-
-select *
-from {{ ref('my_first_dbt_model') }}
-where id = 1
